k8s-deploy/requirements.txt
PyYaml==6.0
